+++ src/entrances/WoodThing.cc
@@ -1,4 +1,4 @@
-//  $Id: WoodThing.cc,v 1.12 2001/04/10 19:42:58 grumbel Exp $
+//  $Id: WoodThing.cc,v 1.13 2001/08/04 12:46:23 grumbel Exp $
 //
 //  Pingus - A free Lemmings clone
 //  Copyright (C) 1999 Ingo Ruhnke <grumbel@gmx.de>
@@ -51,8 +51,8 @@ WoodThing::update(float delta)
     {
       // This does not work and I have no idea why?!
       world->get_particle_holder()
-	->add_particle(new SmokeParticle(pos.x - (surface.get_width()/2) - 24,
-					 pos.y - surface.get_height() + 32 - 147,
+	->add_particle(new SmokeParticle(int(pos.x - (surface.get_width()/2) - 24),
+					 int(pos.y - surface.get_height() + 32 - 147),
 					 -0.6 * (frand() + 1), -0.6 * (frand() + 1)));
       // particle->add_particle(new SmokeParticle());
     }
@@ -63,16 +63,16 @@ WoodThing::draw_offset(int x, int y, float s)
 {
   if (s == 1.0) {
 
-    surface.put_screen(pos.x - (surface.get_width()/2) + x,
-			pos.y - surface.get_height() + 32 + y,
-			counter);
+    surface.put_screen(int(pos.x - (surface.get_width()/2) + x),
+		       int(pos.y - surface.get_height() + 32 + y),
+		       counter);
     //    surface2->put_screen(x_pos - (surface.get_width()/2) + x - 24,
      		 //			y_pos - surface.get_height() + 32 - 147 + y);
 
   } else {
     surface.put_screen((int)((pos.x-32 + x) * s),
-			(int)((pos.y-16 + y) * s),
-			s, s);  
+		       (int)((pos.y-16 + y) * s),
+		       s, s);  
   }
 }
